requirements.txt
requests
pytest
mmpy-bot
redis
python-dotenv
pytest-mock